Optical Character Recognition/requirement.txt
Pillow==3.1.2
pytesseract==0.2.0